[SymForce] return size_t instead of int32_t from values::NumEntries()
Topic: chaoqu_symforce_values_num_entries
Reviewers: aaron, brad, nathan, hayk
GitOrigin-RevId: 527716e9b97b3fe89d0db8362686e46304c58b6b

+++ symforce/opt/values.cc
@@ -65,7 +65,7 @@ void Values<Scalar>::UpdateOrSet(const index_t& index, const Values<Scalar>& oth
 }
 
 template <typename Scalar>
-int32_t Values<Scalar>::NumEntries() const {
+size_t Values<Scalar>::NumEntries() const {
   return map_.size();
 }
